+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Removes the deprecated `create_list` tracker endpoint function as it is no longer available via API
+
 ## v7.4.2 (2024-08-16)
 
 - Fixes an issue with how query params were sent when using the `addCreditCard` function